add data update function and move ssbo source in new file
