+++ demo/2d_colliders/CMakeLists.txt
@@ -15,6 +15,7 @@ set(SOURCE_FILES
     ../../src/rect.cpp
     ../../src/buffers.cpp
     ../../src/pipeline.cpp
+    ../../src/ssbo.cpp
 )
 
 find_package(Vulkan REQUIRED)
