+++ include/HammerEngine/HammerEngine.h
@@ -483,92 +483,20 @@ private:
     VkDescriptorSet SsboDescriptorSet = VK_NULL_HANDLE;
 
 public:
-    // Simple constructor: Just give it the engine instance, data, and its size
-    HammerSSBO(HammerEngine* engine, const void* data, VkDeviceSize size) 
-        : engine(engine), bufferSize(size) {
-        
-        createStorageBuffer(data);
-        allocateAndWriteDescriptorSet();
-    }
+    HammerSSBO(HammerEngine* engine, const void* data, VkDeviceSize size);
 
-    ~HammerSSBO() {
-        if (buffer != VK_NULL_HANDLE) {
-            vkDestroyBuffer(engine->device, buffer, nullptr);
-        }
-        if (bufferMemory != VK_NULL_HANDLE) {
-            vkFreeMemory(engine->device, bufferMemory, nullptr);
-        }
-    }
+    ~HammerSSBO();
 
-    VkBuffer getBuffer() const { return buffer; }
-    VkDeviceSize getSize() const { return bufferSize; }
-    VkDescriptorSet getDescriptorSet() const { return SsboDescriptorSet; }
+    VkBuffer getBuffer();
+    VkDeviceSize getSize();
+    VkDescriptorSet getDescriptorSet();
+
+    void updateData(const void* newData, VkDeviceSize newSize);
 
 private:
-    void createStorageBuffer(const void* data) {
-        VkBufferCreateInfo bufferInfo{};
-        bufferInfo.sType = VK_STRUCTURE_TYPE_BUFFER_CREATE_INFO;
-        bufferInfo.size = bufferSize;
-        bufferInfo.usage = VK_BUFFER_USAGE_STORAGE_BUFFER_BIT;
-        bufferInfo.sharingMode = VK_SHARING_MODE_EXCLUSIVE;
+    void createStorageBuffer(const void* data);
 
-        if (vkCreateBuffer(engine->device, &bufferInfo, nullptr, &buffer) != VK_SUCCESS) {
-            throw std::runtime_error("Failed to create storage buffer!");
-        }
-
-        VkMemoryRequirements memRequirements;
-        vkGetBufferMemoryRequirements(engine->device, buffer, &memRequirements);
-
-        VkMemoryAllocateInfo allocInfo{};
-        allocInfo.sType = VK_STRUCTURE_TYPE_MEMORY_ALLOCATE_INFO;
-        allocInfo.allocationSize = memRequirements.size;
-        allocInfo.memoryTypeIndex = engine->findMemoryType(
-            memRequirements.memoryTypeBits, 
-            VK_MEMORY_PROPERTY_HOST_VISIBLE_BIT | VK_MEMORY_PROPERTY_HOST_COHERENT_BIT
-        );
-
-        if (vkAllocateMemory(engine->device, &allocInfo, nullptr, &bufferMemory) != VK_SUCCESS) {
-            throw std::runtime_error("Failed to allocate storage buffer memory!");
-        }
-
-        vkBindBufferMemory(engine->device, buffer, bufferMemory, 0);
-
-        if (data != nullptr) {
-            void* mappedData;
-            vkMapMemory(engine->device, bufferMemory, 0, bufferSize, 0, &mappedData);
-            memcpy(mappedData, data, static_cast<size_t>(bufferSize));
-            vkUnmapMemory(engine->device, bufferMemory);
-        }
-    }
-
-    void allocateAndWriteDescriptorSet() {
-        VkDescriptorSetAllocateInfo allocInfo{};
-        allocInfo.sType = VK_STRUCTURE_TYPE_DESCRIPTOR_SET_ALLOCATE_INFO;
-        // Accessing engine fields directly:
-        allocInfo.descriptorPool = engine->descriptorPool; 
-        allocInfo.descriptorSetCount = 1;
-        allocInfo.pSetLayouts = &engine->ssboSetLayout;
-
-        if (vkAllocateDescriptorSets(engine->device, &allocInfo, &SsboDescriptorSet) != VK_SUCCESS) {
-            throw std::runtime_error("Failed to allocate SSBO descriptor set!");
-        }
-
-        VkDescriptorBufferInfo bufferInfo{};
-        bufferInfo.buffer = buffer;
-        bufferInfo.offset = 0;
-        bufferInfo.range = bufferSize;
-
-        VkWriteDescriptorSet descriptorWrite{};
-        descriptorWrite.sType = VK_STRUCTURE_TYPE_WRITE_DESCRIPTOR_SET;
-        descriptorWrite.dstSet = SsboDescriptorSet;
-        descriptorWrite.dstBinding = 0; 
-        descriptorWrite.dstArrayElement = 0;
-        descriptorWrite.descriptorType = VK_DESCRIPTOR_TYPE_STORAGE_BUFFER;
-        descriptorWrite.descriptorCount = 1;
-        descriptorWrite.pBufferInfo = &bufferInfo;
-
-        vkUpdateDescriptorSets(engine->device, 1, &descriptorWrite, 0, nullptr);
-    }
+    void allocateAndWriteDescriptorSet();
 };
 
 #endif
